Cadastro e telas de Genero e Plataformas funcionando

+++ recgame/src/main/java/com/br/recgame/repository/PlataformaRepository.java
@@ -4,5 +4,9 @@ import com.br.recgame.model.Plataforma;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface PlataformaRepository extends MongoRepository<Plataforma, Integer> {
+
+public interface PlataformaRepository extends MongoRepository<Plataforma, Long> {
+
 }
+
+
